# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:8080
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /news:
    x-swagger-router-controller: news
    get:
      tags:
        - "get news"
      description: get the news list
      operationId: getNews
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/NewsListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      tags:
        - "news post"
      description: add a news to the list
      operationId: postNews
      parameters:
        - in: header
          name: token
          description: event token
          type: string
          required: true
        - in: body
          name: news
          description: News to create.
          schema:
            $ref: "#/definitions/News"
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/GeneralObjectResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /allnews:
    x-swagger-router-controller: news
    get:
      description: get all news list
      operationId: getAllNews
      tags:
        - "get news"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/NewsListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /news/{id}:
    x-swagger-router-controller: news
    get:
      description: get a news
      operationId: getNewsId
      parameters: 
      - name: id
        type: number
        in: path
        required: true
      responses:
        "200":
          description: Success
          schema: 
           $ref: "#/definitions/GeneralObjectResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: update a news
      # define the parameters
      operationId: updateNews
      tags:
        - "update News"
      parameters:
        - name: id
          description: news id
          type: number
          in: path
          required: true
        - in: header
          name: token
          description: event token
          type: string
          required: true
        - in: body
          name: news
          description: The news to update.
          schema:
            $ref: "#/definitions/News"
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/GeneralObjectResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"  
    delete:
      description: delete a news
      operationId: deleteNewsId
      tags:
        - "delete News"
      # define the parameters
      parameters:
        - name: id
          description: news id
          type: number
          in: path
          required: true
        - in: header
          name: token
          description: event token
          type: string
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralDeleteResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /updateNewsState/{id}:
    x-swagger-router-controller: news
    put:
      tags:
        - "update News"
      description: Update state of news
      operationId: updateStateNews
      parameters:
        - name: id 
          description: news id 
          type: string
          in: path
          required: true
        - in: header
          name: token
          description: news token
          type: string
          required: true
        - in: body 
          name: state
          schema:
            $ref: "#/definitions/State"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralObjectResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /allOrganizations:
    x-swagger-router-controller: organization
    get:
      description: get the organizations list
      operationId: getAllOrganizations
      tags:
        - "organizations"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/OrganizationsListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /organizations:
    x-swagger-router-controller: organization
    get:
      description: get the organizations list
      operationId: getOrganizations
      tags:
        - "organizations"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/OrganizationsListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: add a new organization to the list
      operationId: postOrganization
      tags:
        - "organizations"
      parameters:
        - in: header
          name: token
          description: event token
          type: string
          required: true
        - in: body
          name: organization
          description: The organization to create.
          schema:
            $ref: "#/definitions/Organization"
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/GeneralObjectResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /organizations/{id}:
    x-swagger-router-controller: organization
    get:
      description: get an organization
      operationId: getOrganization
      tags:
        - "organizations"
      parameters: 
      - name: id
        type: number
        in: path
        required: true
      responses:
        "200":
          description: Success
          schema: 
           $ref: "#/definitions/GeneralObjectResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: update an organization
      operationId: updateOrganization
      tags:
        - "organizations"
      parameters:
        - name: id
          description: organization id
          type: number
          in: path
          required: true
        - in: header
          name: token
          description: event token
          type: string
          required: true
        - in: body
          name: organization
          description: The organization to update.
          schema:
            $ref: "#/definitions/Organization"
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/GeneralObjectResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /updateOrganizationState/{id}:
    x-swagger-router-controller: organization
    put:
      tags:
        - "organizations"
      description: Update state of organization
      operationId: updateStateOrganization
      parameters:
        - name: id 
          description: organization id 
          type: string
          in: path
          required: true
        - in: header
          name: token
          description: organization token
          type: string
          required: true
        - in: body 
          name: state
          schema:
            $ref: "#/definitions/State"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralObjectResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  GeneralDeleteResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
        description: a short comment
    required:
      - success
      - description
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  State:
    type: object
    properties:
      state:
        type: boolean
...
  GeneralObjectResponse:
    type: object
    properties:
      id:
        type: number
        description: identifier
    required:
      - id
...
  News:
    type: object
    properties:
      title:
        type: string
        description: Title of the news
      description:
        type: string
        description: Description of the news
      date:
        type: string
        description: Date of news publication
      source:
        type: string
        description: Source of the news
      state:
        type: boolean
        description: State of the event (published, not published)
      image_owner:
        type: string
        description: Image owner of the event
      source_link:
        type: string
        description: Source link of the news
      image:
        type: string
        format: byte
    required:
      - title
      - description
      - date
      - source
      - state
      - image_owner
      - source_link
      - image     
  NewsListResponse:
    required:
      - news
    properties:
      news:
        type: array
        items:
          $ref: "#/definitions/GeneralObjectResponse"
  Organization:
    type: object
    properties:
      name:
        type: string
        description: Name of the organization
      description:
        type: string
        description: Description of the organization
      website:
        type: string
        description: Website of the organization
      address:
        type: string
        description: Address of the organization
      email:
        type: string
        description: Email of the organization
      phone:
        type: string
        description: Phone of the organization
      state:
        type: boolean
        description: indicates if the organization is 'noted' or no
      image:
        type: string
        format: byte
    required:
      - name
      - description 
      - state
      - image
      - email
      - phone
  OrganizationsListResponse:
    required:
      - organizations
    properties:
      organizations:
        type: array
        items:
          $ref: "#/definitions/GeneralObjectResponse"
...